+++ src/outputs.h
@@ -4,8 +4,8 @@
 #include <Arduino.h>
 
 void setupOutputs();
-void setBrightness(uint8_t value);
-uint8_t getCurrentBrightness();
+void setBrightness(uint16_t value);
+uint16_t getCurrentBrightness();
 void setLightOn(bool on);
 bool isLightOn();
 void setModeLed(bool on);
